pyproject.toml
[tool.poetry]
name = "sample_sp"
version = "0.1.0"
description = ""
authors = ["InbalZelinger <inbal@twingate.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
pysaml2 = "^7.1.2"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
